Finish

+++ Model/SaveData.cs
@@ -8,10 +8,10 @@ namespace EnterTarget
         public static DataTable dtTemp1 { get; set; }
         public static DataTable dtTemp2 { get; set; }
         public static string AllModel { get; set; }  
-        public static double op { get; set; }
-        public static double scrap { get; set; }
+        public static int op { get; set; }
+        public static int scrap { get; set; }
         public static string Model { get; set; }
-        public static double opTarget { get; set; }
-        public static double scrapTarget { get; set; }
+        public static int opTarget { get; set; }
+        public static int scrapTarget { get; set; }
     }
 }
